Use 'INTEGER PRIMARY KEY' for ID columns

+++ migrations/2019-01-26-013326_create_scribbles_tags_taggings/up.sql
@@ -1,17 +1,17 @@
 CREATE TABLE scribbles (
-    id         BIGINT PRIMARY KEY NOT NULL,
+    id         INTEGER PRIMARY KEY NOT NULL,
     created_at BIGINT NOT NULL,
     text       TEXT NOT NULL
 );
 
 CREATE TABLE tags (
-    id         BIGINT PRIMARY KEY NOT NULL,
+    id         INTEGER PRIMARY KEY NOT NULL,
     created_at BIGINT NOT NULL,
     text       TEXT UNIQUE NOT NULL
 );
 
 CREATE TABLE taggings (
-    id          BIGINT PRIMARY KEY NOT NULL,
+    id          INTEGER PRIMARY KEY NOT NULL,
     created_at  BIGINT NOT NULL,
     scribble_id BIGINT NOT NULL,
     tag_id      BIGINT NOT NULL
